requirements.txt
aiodns==3.1.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
# binance-futures-connector==4.0.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
dateparser==1.1.8
frozendict==2.3.8
frozenlist==1.4.0
html5lib==1.1
idna==3.4
lxml==4.9.3
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.0
pandas==2.1.0
pycares==4.4.0
pycparser==2.21
pycryptodome==3.19.0
# python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
six==1.16.0
soupsieve==2.5
tradingview-ta==3.3.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.5
webencodings==0.5.1
yarl==1.9.2
# yfinance==0.2.28
